查询任意桩号版本

+++ src/main/java/com/ljj/flinkquery/demos/entity/VehicleSeg.java
@@ -18,5 +18,5 @@ public class VehicleSeg {
     private Integer vehicleType;
     private String specialFlag;
     private int laneNo;
-    private double averageSpeed;
+    private float averageSpeed;
 }
